Adjust pie chart offset to avoid overlapping labels
To reduce the change of overlapping labels, the pie chart's `startOffset`
defaults to `Math.PI / 4` so that the smaller labels are likely to appear
towards the right side of the chart, where there is more space.

Also, redraw the chart when `startOffset` changes. This makes it easy to
specify and change the rotation of the chart.